Remove a service from connman when deleting a profile
:Release Notes:
Call the connman API to remove a service so that its security information
like passphrase gets deleted from connman.

:Detailed Notes:
With just the disconnect call, connman disconnects a service , but saves
its passphrase, so that next time when connecting to this same service, no
passphrase information is asked. From the UI perspective this means a user
is able to connect even with an incorrect passphrase.
So this fix adds the API for removing a service, which deletes all the
security information about this service, thereby not letting any user
to connect again with incorrect passphrase.

:Testing performed:
Called the APIs for "connect" and "deleteprofile" one after the other and
confirmed that incorrect passphrase is rejected.

:Issues Addressed:
[GF-8089] wifi service: Disconnect from network doesn't work.

Open-webOS-DCO-1.0-Signed-off-by: Sapna Todwal <sapna.todwal@lge.com>

Change-Id: I3531a2306e7674fda03f5ba3833832fd50abca76
Reviewed-on: https://g2g.palm.com/2910
Reviewed-by: Sapna Todwal <sapna.todwal@lge.com>
Tested-by: Sapna Todwal <sapna.todwal@lge.com>