Handle "services_added" and "services_removed" arguments separately
:Release Notes:
Instead of calling the functions "connman_manager_update_services" and
"connman_manager_remove_old_services" in an OR statement, evaluate the
functions separately.

:Detailed Notes:
Since the two functions were called in an OR statement, if first one
returned TRUE , second one was never called, resulting in cases where
if both "services_added" and "services_removed" arguments were valid,
only "services_added" was handled and "services_removed" was ignored.
With this fix, now both the arguments will be handled one by one.

:Testing Performed:
The case of both "services_added" and "services_removed" being non-NULL
can be easily reproduce if the security type for an AP is changed, since
then the network entry for this same AP with old security type is removed
and the entry for the same AP with new network type is added.
So this case was tested and it was confirmed that with this fix only one
network entry for this AP is seen with the correct security type.

:Issues addressed:
[GF-9968] Seeing 2 networks instead of one when the security for the
          corresponding AP is changed

Open-webOS-DCO-1.0-Signed-off-by: Sapna Todwal <sapna.todwal@lge.com>


Change-Id: Ib4c0e33fa0b597a83343cae950f37e54aee3e20c
Reviewed-on: https://g2g.palm.com/2912
Reviewed-by: Sapna Todwal <sapna.todwal@lge.com>
Tested-by: Sapna Todwal <sapna.todwal@lge.com>