Update technology properties on receiving "PropertyChanged" signal
:Release Notes:
Update the technology properties such as "Powered" and "P2P" on
receiving the "PropertyChanged" signal

:Detailed Notes:
In many cases the "Powered" property for a technology is updated
later, i.e after the technology is added. This case is different
from the property being updated by the "setstate" API. So this fix
resolved the issue were there was a mismatch between the "powered"
field for a technology in webos-connman-adapter and the "Powered"
state for a technology in connman.

:Testing Performed:
To reproduce the issue described in the jira, only thing required
was to restart the connman service. Once that was done the "powered"
state for wifi technology was FALSE, and the "findnetworks" API reported
"WiFi switched off", although it was actually powered on.
With this fix, even when connman was restarted , the "powered" field
for wifi technology was always in sync with the value in connman, and
"findnetworks" API returned success.

:Issues addresses:
[GF-10268] Webos-connman-adapter does not work properly if it was started
           before connman

Open-webOS-DCO-1.0-Signed-off-by: Sapna Todwal <sapna.todwal@lge.com>

Change-Id: Ief8281ff3c64dab4b03bdfc01da31e32b65a7e7a
Reviewed-on: https://g2g.palm.com/2913
Reviewed-by: Sapna Todwal <sapna.todwal@lge.com>
Reviewed-by: Build Verification
Tested-by: Sapna Todwal <sapna.todwal@lge.com>