Implement reliable wifi scan behaviour
:Release Notes:
Scanning for WiFi networks is only executed as long as we have users
subscribed for it.

:Detailed Notes:
connman requires the user to ask for fresh scan results whenever the
UI should be updated. As we're sitting below the UI and the UI is
subscribed to us to be informed about new scan results we have to take
the job of rescheduling a new scan request in specific interval as
long as we have any clients wanting fresh scan results. This is
implemented through counting the number of current ls2 subscriptions
and is done after every timeout before a new scan should be scheduled.

As sitenote: This has nothing to do with the background scaning feature
connman offers. The autoscan functionality is only meant to be used for
fast auto connection on device start. For roaming the bgscan functionality
from wpa-supplicant is used to monitor the ssid the device is currently
connected to for any changes.

:Testing Performed:
1. Compiled and installed webos-connman-adapter on the M13 board
2. Opened the settings application and navigated to the network settings
3. Activated wireless networking
3. Waited for the listed of available networks populated with some entries
4. Verified that the list stays stable over some time (~30min)
5. Connected to a available secured network
6. Closed network settings and reopened it
7. Verified the list of networks is correct and stays stable over some time
   (~30min)

:QA Notes:
When testing this change please respect the following:
- connecting to wifi networks doesn't work reliable on H13 and will
  cause inconsistencies in the service states so it will happen
  directly after the connection failed that the list of networks is
  getting cleared. It should be filled again after 30 seconds and
  should be filled until you tried to connect again and the connection
  fails again.
- when you flash a new image on your TV make sure you do a factory
  reset of the TV so /var/lib/connman gets removed entirely.

:Issues Addressed:
[GF-13208] Rework WiFi scan behaviour inside webos-connman-adapter

Open-webOS-DCO-1.0-Signed-off-by: Simon Busch <simon.busch@lge.com>

Change-Id: I7158c6971f36bf85b7684786371385b4fe5f238d
Reviewed-on: https://g2g.palm.com/3676
Reviewed-by: Build Verification
Reviewed-by: Sapna Todwal <sapna.todwal@lge.com>
Tested-by: Sapna Todwal <sapna.todwal@lge.com>