Use higher timeout for dbus calls
:Release Notes:
Use a higher timeout for dbus calls should make API calls more reliable.

:Detailed Notes:
We discovered that sometimes it takes longer to connect to a service than
the configured timeout. This can be related to different things (hardware,
driver, ...) so we decided to configure a higher timeout for the related
dbus calls. We're setting the timeout for all dbus calls to the service
object as it's only possible to configured it per proxy and not per call.

:Testing Performed:
Testing this isn't really possible as it depends on different aspects
which are not reliable reproducible.

:QA Notes:

:Issues Addressed:
[GF-16341] Set higher timeout for dbus calls to connman

Open-webOS-DCO-1.0-Signed-off-by: Simon Busch <simon.busch@lge.com>

Change-Id: Idaf0cb317c4794c4b477009adeb4799d1bf23052
Reviewed-on: https://g2g.palm.com/3678
Reviewed-by: Build Verification
Reviewed-by: Sapna Todwal <sapna.todwal@lge.com>
Tested-by: Sapna Todwal <sapna.todwal@lge.com>